# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCDMB API Documentation",
        "description": "Comprehensive API documentation for the NCDMB Document Management System. This API provides endpoints for document management, workflow automation, procurement, legal processes, and more.",
        "contact": {
            "email": "support@ncdmb.gov.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/legal/reviews": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "List all legal reviews",
                "operationId": "d347eb213fc3640c48747fbf767b6eeb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Legal"
                ],
                "summary": "Create a new legal review",
                "operationId": "90e65897491da609c6ee64f02c3ce596",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LegalReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Legal review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/reviews/{id}": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "Get a specific legal review",
                "operationId": "0fd0dfe4976dbcd838ad75550b919315",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Legal"
                ],
                "summary": "Update a legal review",
                "operationId": "6bf28e2cd5cebc5854e8b7bc8cbaae37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LegalReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Legal review updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Legal"
                ],
                "summary": "Delete a legal review",
                "operationId": "aea1b6a0b4891a224127ccd5e61413c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legal review deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/reviews/{id}/approve": {
            "post": {
                "tags": [
                    "Legal"
                ],
                "summary": "Approve a legal review",
                "operationId": "30385debf4bf2f16aa77d0d640063060",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legal review approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/reviews/{id}/reject": {
            "post": {
                "tags": [
                    "Legal"
                ],
                "summary": "Reject a legal review",
                "operationId": "3e9de24aa090ff0f1bf6c7d6fe46877d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legal review rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/reviews/project/{projectId}": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "Get reviews for a specific project",
                "operationId": "9b95afcdfbfa93a18e957e6943cf4d57",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/reviews/contract/{contractId}": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "Get reviews for a specific contract",
                "operationId": "09cc7c1d3265d6cd5e3a7ecd6e055291",
                "parameters": [
                    {
                        "name": "contractId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Success": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    }
                },
                "type": "object"
            },
...
            "LegalReviewRequest": {
                "required": [
                    "review_type",
                    "reviewed_by"
                ],
                "properties": {
                    "project_contract_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "document_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "review_type": {
                        "type": "string",
                        "enum": [
                            "contract_review",
                            "compliance_check",
                            "risk_assessment",
                            "variation_review",
                            "termination_review",
                            "other"
                        ],
                        "example": "contract_review"
                    },
                    "reviewed_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "review_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_review",
                            "approved",
                            "rejected",
                            "conditional"
                        ],
                        "example": "pending",
                        "nullable": true
                    },
                    "review_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15",
                        "nullable": true
                    },
                    "legal_opinion": {
                        "type": "string",
                        "example": "Legal opinion text",
                        "nullable": true
                    },
                    "compliance_score": {
                        "type": "number",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 85.5,
                        "nullable": true
                    },
                    "risks_identified": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "recommendations": {
                        "type": "string",
                        "example": "Recommendations",
                        "nullable": true
                    },
                    "requires_revision": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "revision_notes": {
                        "type": "string",
                        "example": "Revision notes",
                        "nullable": true
                    },
                    "approved_by": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "approval_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-20",
                        "nullable": true
                    },
                    "rejection_reason": {
                        "type": "string",
                        "example": "Rejection reason",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication. Enter token in format: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}